backend/requirements.txt
Django==5.2.4
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.31.0
django-environ==0.11.2
djangorestframework==3.14.0
dj-database-url==3.0.1

# Autenticación JWT - Agregado 9/07/25
djangorestframework-simplejwt==5.3.1

# 🆕 NUEVAS DEPENDENCIAS - Versión 2.0.0
# Manejo de imágenes y archivos multimedia
Pillow==10.4.0
# Validación y procesamiento de datos
django-filter==24.3
# Utilidades para desarrollo
django-extensions==3.2.3

# 🚀 DEPENDENCIAS PARA PRODUCCIÓN
# Servidor WSGI para producción
gunicorn==21.2.0
# Manejo de archivos estáticos en producción
whitenoise==6.6.0

